Initial version of pip installable version
