+++ pyproject.toml
@@ -1,0 +1,41 @@
+[project]
+name = "astroOracle"
+version = "0.1.0"
+description = "A package for real time, hierarchical classification of LSST alerts"
+authors = [
+  { name="Ved Shah"},
+]
+readme = "README.md"
+dependencies = [
+    "astropy==6.1.0",
+    "graphviz==0.20.3",
+    "keras==2.15.0",
+    "matplotlib==3.9.0",
+    "networkx==3.3",
+    "numpy==1.26.4",
+    "pandas==2.2.2",
+    "polars==0.20.31",
+    "scikit-learn==1.5.0",
+    "scipy==1.13.1",
+    "seaborn==0.13.2",
+    "tensorflow==2.15.0",
+    "tqdm==4.66.4",
+    "umap-learn==0.5.7"
+]
+
+[build-system]
+requires = ["setuptools>=42"]
+build-backend = "setuptools.build_meta"
+
+[tool.setuptools.packages.find]
+where = ["src"]
+
+[project.scripts]
+Oracle-train = "astroOracle.train_RNN:main"
+Oracle-test = "astroOracle.train_RNN:main"
+Oracle-runAnalysis = "astroOracle.run_analysis:main"
+Oracle-classSummaries = "astroOracle.class_summaries:main"
+Oracle-combineParquet = "astroOracle.combine_parquet:main"
+Oracle-fitsToParquet = "astroOracle.fits_to_parquet:main"
+Oracle-prepArrays = "astroOracle.prep_arrays:main"
+Oracle-timeBenchmark = "astroOracle.time_benchmark:main"
